v0.4.1 Prototype 1
This adds dialogue to the New Hope Central, and sets the foundation for the prospector group in this region.

They will either engage in combat if the prospectors encountered at the crash site were killed, or not if they survived.

Either way, they will drop data on where to go next.