Update to version v0.4.1
New Hope has been finished off and as expected, this project is going smoothly now that all the framework is in place.

New NPCs have been added, and you can now see the effects of prior choices with the group of prospectors there. This will be the last of the prospectors, that much I can promise.